+++ features/src/commonMain/kotlin/io/github/alelk/pws/features/tags/tagsScreenModule.kt
@@ -6,5 +6,5 @@ import io.github.alelk.pws.features.tags.songs.TagSongsScreen
 
 val tagsScreenModule = screenModule {
   register<SharedScreens.Tags> { TagsScreen() }
-  register<SharedScreens.TagSongs> { provider -> TagSongsScreen(provider.tagId) }
+  register<SharedScreens.TagSongs> { provider -> TagSongsScreen(provider.tagIdString) }
 }